Added repository outlineview, revamped TableViewController heirachy
I added StandardTableViewCell to simly hold the text fields and icons. FileSystemTableViewCell is now a sublcass of StandardTableViewCell and just adds configurations to the labels and icons. This commit also contains dummy classes for repos, branches, and containers that needs some work to optimise code before Nanashili can connect his code to mine.