Added proportional margins and auto-collapsing of disabled hitboxes
if a certain drag and drop action is disabled (eg. leading and trailing for the navigation sidebar), the hitboxes for those drags no longer take up space, leaving more hitbox area for the other possible drag actions.

Proportional margins allow the hitboxes to be a relative size to the rectangle containing it.