Improved about screen (#615)
## Summary

<!--- REQUIRED: Provide a clear and concise summary of what this PR aims
to achieve. -->

## Changes Made
Refactor: Eliminated repetitive code in the `loadCredits` and
`loadLicense` functions by creating a generic function to read the
contents of a file, improving code maintainability and reducing
redundancy.

Extracted window configuration to a configureWindow method for improved
readability.

Introduced a setupEscapeKeyEventHandler method to centralize event
monitor setup.

Created an animateWindow method to handle window animations, reducing
code duplication.

Added a removeEscapeKeyEventHandler method to ensure the event monitor
is removed during deinitialization.

Added a required init?(coder:) method to conform to NSWindowController
subclass requirements, even though it throws an error because the class
is not intended to be initialized from a coder.

Added support for cache images and improved handling of about states,
enhancing the user experience and visual presentation.

<!--- REQUIRED: Describe the changes introduced by this PR in a clear
and concise manner. -->

## TODO
NONE
<!--- REQUIRED: Outline any pending tasks or items that need further
attention within this PR. -->

## Motivation
The user experience with the About screen was not very performant and
the constant image loading was not ideal
<!--- REQUIRED: Explain the primary goal and purpose of this PR. Why is
it necessary, and how does it benefit the project? -->

## Testing
Opened and closed multiple times
<!--- REQUIRED: Detail the testing process you have undertaken,
including what aspects of the code or functionality you tested. -->

## Screenshots/GIFs
NONE
<!--- If applicable, include screenshots or GIFs that visually
demonstrate the changes made in this PR. -->

## Checklist

Please ensure all of the following are completed before submitting the
PR:

- [X] Code has been tested and verified.
- [X] Documentation has been updated to reflect the changes.
- [X] All tests pass successfully.
- [X] Code follows the project's coding guidelines and style.
- [X] The branch is up-to-date with the latest changes from the main
branch.

## Related Issues

This PR addresses the following issue(s): NONE

## Additional Notes
NONE
<!--- REQUIRED: If you have any additional notes, considerations, or
context to add regarding this PR, please do so here. -->

---------

Co-authored-by: Wesley de Groot <email@wesleydegroot.nl>