Feature: Added the new Aurora Editor to the Editor (#633)
## Summary

This PR adds support for the new Editor in AuroraEditor

## Changes Made
Removed the old editor in favour for the new one
<!--- REQUIRED: Describe the changes introduced by this PR in a clear
and concise manner. -->

## TODO
Need to remove the old editor code still
<!--- REQUIRED: Outline any pending tasks or items that need further
attention within this PR. -->

## Motivation
We needed the new editor duh
<!--- REQUIRED: Explain the primary goal and purpose of this PR. Why is
it necessary, and how does it benefit the project? -->

## Testing
Basic file highlighting tests
<!--- REQUIRED: Detail the testing process you have undertaken,
including what aspects of the code or functionality you tested. -->

## Screenshots/GIFs

![image](https://github.com/AuroraEditor/AuroraEditor/assets/63672227/7ae72a8d-57d5-4d7e-9a50-eedd0d90a3d7)

<!--- If applicable, include screenshots or GIFs that visually
demonstrate the changes made in this PR. -->

## Checklist

Please ensure all of the following are completed before submitting the
PR:

- [X] Code has been tested and verified.
- [X] Documentation has been updated to reflect the changes.
- [X] All tests pass successfully.
- [X] Code follows the project's coding guidelines and style.
- [X] The branch is up-to-date with the latest changes from the main
branch.
- [X] Reviewed by at least one other contributor.

## Related Issues

This PR addresses the following issue(s): [<!--- REQUIRED: Issue
Number-->] NONE

## Additional Notes
NONE
<!--- REQUIRED: If you have any additional notes, considerations, or
context to add regarding this PR, please do so here. -->

---------

Signed-off-by: Nanashi Li <nanashili@plexus.dev>