Cleanup: Remove unwanted CodeEdit credit (#635)
## Summary
This removes all credits to CodeEdit in the files.
<!--- REQUIRED: Provide a clear and concise summary of what this PR aims
to achieve. -->

## Changes Made
Removes all credits to CodeEdit in the files.
<!--- REQUIRED: Describe the changes introduced by this PR in a clear
and concise manner. -->

## TODO
None
<!--- REQUIRED: Outline any pending tasks or items that need further
attention within this PR. -->

## Motivation
We do not need to add these to files since our ReadMe clearly states we
were a fork of CE.
<!--- REQUIRED: Explain the primary goal and purpose of this PR. Why is
it necessary, and how does it benefit the project? -->

## Testing
NONE
<!--- REQUIRED: Detail the testing process you have undertaken,
including what aspects of the code or functionality you tested. -->

## Screenshots/GIFs
NONE
<!--- If applicable, include screenshots or GIFs that visually
demonstrate the changes made in this PR. -->

## Checklist

Please ensure all of the following are completed before submitting the
PR:

- [X] Code has been tested and verified.
- [X] Documentation has been updated to reflect the changes.
- [X] All tests pass successfully.
- [X] Code follows the project's coding guidelines and style.
- [X] The branch is up-to-date with the latest changes from the main
branch.
- [X] Reviewed by at least one other contributor.

## Related Issues

This PR addresses the following issue(s): [<!--- REQUIRED: Issue
Number-->] NONE

## Additional Notes
NONE
<!--- REQUIRED: If you have any additional notes, considerations, or
context to add regarding this PR, please do so here. -->